% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model}
\alias{model}
\title{Exercises on statistical modelling in R}
\format{yaml-file with :
\describe{
  \item{id}{unique id for each exercise}
  \item{domain}{self-defined domain that exercise belongs to (affects the edu-package addin)}
  \item{grade}{self-defined difficulty of the exercise. for instance easy, moderate or hard}
  \item{lang_en}{Question in English}
  \item{lang_fi}{Question in Finnish or any other language}
  \item{extra}{Some extra preformatted content you want to be displayed with the exercise}
  \item{ans}{Correct answer or multiple alternative answers}
}}
\source{
\url{https://github.com/ropengov/edudata}
}
\usage{
model
}
\description{
This data provides exercises for statistical modelling in R for edu-package
}
\keyword{datasets}
